airflow-concepts-2/requirements.txt
pandas
scikit-learn
joblib
apache-airflow

